+++ wiki/ReleaseNotes.md
@@ -14,6 +14,10 @@ None.
 * **[C++]** Cluster controller now uses the new `touch` and `removeStale` methods of
   [`GlobalCapabilitiesDirectory.fidl`](../basemodel/src/main/franca/joynr/GlobalCapabilitiesDirectory.fidl).
   The Global Capabilities Directory has to implement version 0.3 of the interface.
+* **[Android]** Added support for multi-user in the Android Binder runtime, allowing Android
+  Services to be bound either as the system user or as a specific user. The implementation is
+  tailored for the CC to run in user 0 (system), and joynr app clients can connect on
+  whichever user they decide to in a system with multi-user capabilities.
 
 # joynr 1.14.0